+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+------------
+
+### Added
+
+- Don't choose a default worksheet on first startup
+
+
 [0.9.1] - 2026-06-28
 --------------------
 
@@ -490,6 +498,7 @@ Phase 1: Core Domain.
 - Numeric literals with decimal point support
 - 373 passing tests
 
+[Unreleased]: https://github.com/wisnij/unitary/compare/v0.9.1...HEAD
 [0.9.1]: https://github.com/wisnij/unitary/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/wisnij/unitary/compare/v0.8.3...v0.9.0
 [0.8.3]: https://github.com/wisnij/unitary/compare/v0.8.2...v0.8.3